src/requirements.txt
Pyro4
pySerial